flask/requirements.txt
flask
python-dotenv
flask-cors
python-dotenv
web3

